using custom weather icons; ic_launcher.png; better weather activity ui

+++ app/src/main/java/com/stedi/weatherapp/model/data/owmweather/Main.kt
@@ -1,3 +1,3 @@
 package com.stedi.weatherapp.model.data.owmweather
 
-data class Main(var temp: Double?, var pressure: Int?, var humidity: Int?)
+data class Main(var temp: Double?, var pressure: Double?, var humidity: Int?)
